+++ pyoffers/models/offer.py
@@ -14,6 +14,9 @@ class Offer(Model):
     def add_target_country(self, country_code):
         return self._manager.add_target_country(self['id'], country_code)
 
+    def get_target_countries(self):
+        return self._manager.get_target_countries(self['id'])
+
     def add_category(self, category_id):
         return self._manager.add_category(self['id'], category_id)
 
@@ -37,6 +40,9 @@ class OfferManager(ModelManager):
 
     def add_target_country(self, id, country_code):
         return self._call('addTargetCountry', id=id, country_code=country_code)
+
+    def get_target_countries(self, id):
+        return self._call('getTargetCountries', id=id, single_result=False)
 
     def add_category(self, id, category_id):
         return self._call('addCategory', id=id, category_id=category_id)